Extend status notice text.
Attempting to use some keywords in bold.
We have many repositories which need some status section like this;
while we don't currently have a standard for this text,
it would be nice to attempt to have one, and I may start pushing
towards that, starting first with just a convention on key words.